add trait
